backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
pillow
requests
python-dotenv
